+++ site_content.json
@@ -17,6 +17,7 @@
     "layout": {
       "navigation": {
         "about": "銭湯ぐらしとは",
+        "media": "読みもの",
         "kosugiyu": "小杉湯",
         "contact": "お問い合わせ"
       },
