+++ site_content.yml
@@ -13,6 +13,7 @@ common:
   layout:
     navigation:
       about: 銭湯ぐらしとは
+      media: 読みもの
       kosugiyu: 小杉湯
       contact: お問い合わせ
     footer:
