--- src/scripts/mediaArticle.js
@@ -1,11 +1,0 @@
-import SimpleParallax from './class/simpleParallax.js';
-
-const $window = $(window);
-const simpleParallax = new SimpleParallax();
-
-$(() => {
-  simpleParallax.addElement('.js-MediaArticle__photoMain');
-  simpleParallax.init();
-});
-
-$window.on('load', mainvisualAnimationSinglePhoto);
