+++ src/styles/article.scss
@@ -6,6 +6,191 @@ $photoMainHeightBase: 650px;
 $photoSubWidthBase: 1280px;
 $photoSubHeightBase: 500px;
 
+@mixin normalLink {
+  border-bottom: dashed 1px #aaa;
+  transition: opacity .4s;
+
+  &:link,
+  &:visited {
+    color: #999;
+  }
+  &:hover {
+    opacity: .6;
+  }
+
+}
+
+.MainVisual {
+  background-position: center center;
+  background-size: cover;
+  position: relative;
+  //padding-bottom: 20%;
+  height: 200px;
+
+  &__title {
+    position: absolute;
+    @include baseHeadingLv1;
+    text-align: center;
+    color: #fff;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, .5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  &__image {
+    // TODO: カルーセル化
+    width: 100%;
+    padding-bottom: 20%;
+  }
+}
+
+.CategoryNavigation {
+  display: flex;
+  margin:30px auto 50px;
+  align-items: center;
+  justify-content: center;
+
+  & > li {
+    &.is-active {
+      border-bottom: solid 1px #000;
+    }
+
+    & > a {
+      display: block;
+      text-align: center;
+      margin: 0 20px;
+      //font-family: serif;
+      font-size: rem(14px);
+      padding: 20px;
+
+      @include mq('sm') {
+        padding: 20px 0;
+      }
+
+      &:link,
+      &:visited {
+        color: #000;
+      }
+    }
+  }
+}
+
+.CellList {
+  width: 100%;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, 320px);
+  justify-content: center;
+
+  @include mq('sm') {
+    grid-template-columns: auto;
+  }
+}
+.Cell {
+  margin-bottom: 40px;
+  transition: .5s;
+
+  &:hover {
+    opacity: .8;
+  }
+
+  @include mq('sm') {
+    padding: 20px;
+  }
+
+  &__thumbNail {
+    width: 300px;
+    padding-bottom: (2 / 3) * 100%; // TODO アドホック
+    background-size: cover;
+    background-position: center;
+    background-repeat: no-repeat;
+
+    @include mq('sm') {
+      width: 100%;
+    }
+  }
+
+  &__main {
+    padding-right: 20px;
+  }
+
+  &__category {
+    display: inline-block;
+    color: #fff;
+    margin: 10px 0 0;
+    padding: 2px 5px;
+    font-size: rem(10px);
+
+    &--event {
+      background-color: #ca6492;
+    }
+
+    &--news {
+      background-color: #40bb8f;
+    }
+
+    &--interview {
+      background-color: #88f;
+    }
+  }
+
+  &__title {
+    margin: 10px 0;
+    font-family: $base-font-yumincho;
+    color: $system-color-text;
+    font-size: rem(22px);
+  }
+
+  &__date {
+    margin: 5px 0;
+    color: #999;
+    font-size: rem(14px);
+    letter-spacing: .2em;
+  }
+
+  &__tags {
+    font-size: rem(12px);
+
+    & > li {
+      display: inline-block;
+      margin-right: 5px;
+      letter-spacing: .2em;
+
+      & > a {
+
+        &:link,
+        &:visited {
+          color: #999;
+        }
+
+        &:before {
+          content: "#";
+        }
+      }
+    }
+  }
+}
+
+@mixin normalLink {
+  //display: inline-block;
+  //paddding-bottom: 2px;
+  border-bottom: dashed 1px #aaa;
+  transition: opacity .4s;
+
+  &:link,
+  &:visited {
+    color: #999;
+  }
+  &:hover {
+    opacity: .6;
+  }
+
+}
+
 .Article {
 
   &__inner {
@@ -17,17 +202,67 @@ $photoSubHeightBase: 500px;
     };
   }
 
-  &__headingWrapper {
-    margin: 40px 0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
+  &__title {
+
+    &Info {
+      text-align: center;
+      vertical-align: middle;
+      letter-spacing: .1em;
+    }
+
+    &Category {
+      display: inline-block;
+      padding: 4px 6px;
+      color: #fff;
+      background-color: #ca6492;
+      //font-size: rem(10px);
+      margin-right: 10px;
+      line-height: 1;
+    }
+
+    &Date {
+      display: inline-block;
+      color: #999;
+    }
+
+    &SeriesName {
+      margin-left: 10px;
+      display: inline-block;
+
+      & > a:link,
+      & > a:visited {
+        @include normalLink;
+      }
+    }
+
+    &Tags {
+      margin-top: 15px;
+      font-size: rem(12px);
+
+      & > li {
+        display: inline-block;
+        margin-right: 2px;
+
+        & > a {
+          @include normalLink;
+
+          &:before {
+            content: "#";
+            color: #999;
+          }
+        }
+
+      }
+    }
   }
+
+
 
   &__headingLv1 {
     @include baseHeadingLv1;
     margin: 50px 0;
-    font-size: rem(60px); // TODO アドホック
+    font-size: rem(45px); // TODO アドホック
+    //text-align: left;
 
     @include mq('sm') {
       margin: 20px 0;
@@ -38,7 +273,7 @@ $photoSubHeightBase: 500px;
   &__headingLv2 {
     margin: 60px 0;
     font-family: $base-font-serif;
-    font-size: rem(34px);
+    font-size: rem(26px);
     color: $system-color-text;
     line-height: 1.4;
 
@@ -48,47 +283,52 @@ $photoSubHeightBase: 500px;
     };
   }
 
+/*
   &__headingLv3 {
     margin: 40px 0 30px;
-    font-weight: 600;
-    font-size: rem(18px);
+    font-size: rem(22px);
     font-family: $base-font-serif;
     color: $system-color-text;
+  }
+*/
 
-    .Article__headingLv2 + & {
-      margin-top: -20px;
-    }
-
-    & + .Article__text  {
-      margin-top: 0px;
-    }
+  &__headingLv3 {
+    margin: 40px 0 30px;
+    font-size: rem(18px);
+    //font-family: $base-font-serif;
+    color: $system-color-text;
   }
 
   &__text {
     @include baseText;
-    margin: 60px 0 90px;
+    margin: 20px 0 60px;
 
     @include mq('sm') {
       margin: 20px 0 60px;
     };
 
-/*
     & + & {
-      margin-top: -70px; // TODO: ここびみょう
-
-      @include mq('sm') {
-        margin-top: 0;
-      };
+      margin-top: -20px;
     }
-*/
+
+    & > a {
+      @include normalLink;
+      line-height: 1.3;
+    }
+
   }
 
   &__innerImage {
-    margin-bottom: 90px;
+    margin-bottom: 60px;
 
     & > img {
       width: 100%;
     }
+  }
+
+  &__caption {
+    font-size: rem(14px);
+    color: #999;
   }
 
   &__photoWrapper {
@@ -105,11 +345,7 @@ $photoSubHeightBase: 500px;
   }
 
   &__photoMain {
-    padding-bottom: percentage($photoMainHeightBase / $photoMainWidthBase);
-
-    &.is-animate {
-      animation: mainvisual-animation 6s linear;
-    }
+    padding-bottom: 30%; // TODO: アドホック
 
     @include mq('sm') {
       padding-bottom: 100%;
@@ -125,19 +361,6 @@ $photoSubHeightBase: 500px;
     @include mq('sm') {
       padding-bottom: 100%;
       transform: scale(1.2);
-    };
-  }
-
-  &__illust {
-    margin-right: 30px;
-    width: 180px;
-    height: 180px;
-    background-size: contain;
-    background-repeat: no-repeat;
-    @include mq('sm') {
-      margin-right: 10px;
-      width: 100px;
-      height: 100px;
     };
   }
