--- src/styles/media-top.scss
@@ -1,176 +1,0 @@
-@import "util";
-@import "foundation/base-style";
-
-$photoMainWidthBase: 1280px;
-$photoMainHeightBase: 650px;
-$photoSubWidthBase: 1280px;
-$photoSubHeightBase: 500px;
-
-@mixin normalLink {
-  border-bottom: dashed 1px #aaa;
-  transition: opacity .4s;
-
-  &:link,
-  &:visited {
-    color: #999;
-  }
-  &:hover {
-    opacity: .6;
-  }
-
-}
-
-.MainVisual {
-  background-position: center center;
-  background-size: cover;
-  position: relative;
-  //padding-bottom: 20%;
-  height: 200px;
-
-  &__title {
-    position: absolute;
-    @include baseHeadingLv1;
-    text-align: center;
-    color: #fff;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, .5);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  &__image {
-    // TODO: カルーセル化
-    width: 100%;
-    padding-bottom: 20%;
-  }
-}
-
-.CategoryNavigation {
-  display: flex;
-  margin:30px auto 50px;
-  align-items: center;
-  justify-content: center;
-
-  & > li {
-    &.is-active {
-      border-bottom: solid 1px #000;
-    }
-
-    & > a {
-      display: block;
-      text-align: center;
-      margin: 0 20px;
-      //font-family: serif;
-      font-size: rem(14px);
-      padding: 20px;
-
-      @include mq('sm') {
-        padding: 20px 0;
-      }
-
-      &:link,
-      &:visited {
-        color: #000;
-      }
-    }
-  }
-}
-
-.CellList {
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, 320px);
-  justify-content: center;
-
-  @include mq('sm') {
-    grid-template-columns: auto;
-  }
-}
-.Cell {
-  margin-bottom: 40px;
-  transition: .5s;
-
-  &:hover {
-    opacity: .8;
-  }
-
-  @include mq('sm') {
-    padding: 20px;
-  }
-
-  &__thumbNail {
-    width: 300px;
-    padding-bottom: (2 / 3) * 100%; // TODO アドホック
-    background-size: cover;
-    background-position: center;
-    background-repeat: no-repeat;
-
-    @include mq('sm') {
-      width: 100%;
-    }
-  }
-
-  &__main {
-    padding-right: 20px;
-  }
-
-  &__category {
-    display: inline-block;
-    color: #fff;
-    margin: 10px 0 0;
-    padding: 2px 5px;
-    font-size: rem(10px);
-
-    &--event {
-      background-color: #ca6492;
-    }
-
-    &--news {
-      background-color: #40bb8f;
-    }
-
-    &--interview {
-      background-color: #88f;
-    }
-  }
-
-  &__title {
-    margin: 10px 0;
-    font-family: $base-font-yumincho;
-    color: $system-color-text;
-    font-size: rem(22px);
-  }
-
-  &__date {
-    margin: 5px 0;
-    color: #999;
-    font-size: rem(14px);
-    letter-spacing: .2em;
-  }
-
-  &__tags {
-    font-size: rem(12px);
-
-    & > li {
-      display: inline-block;
-      margin-right: 5px;
-      letter-spacing: .2em;
-
-      & > a {
-
-        &:link,
-        &:visited {
-          color: #999;
-        }
-
-        &:before {
-          content: "#";
-        }
-      }
-    }
-  }
-}
