+++ src/templates/_layout/_navigation.ejs
@@ -13,6 +13,7 @@
   <a href="/" class="Navigation__sitelogo"><%= sitedata.siteTitle %></a>
   <ul class="Navigation__main">
     <li><a href="/about/project.html"><%= sitedata.common.layout.navigation.about %></a></li>
+    <li><a href="/about/project.html"><%= sitedata.common.layout.navigation.media %></a></li>
   </ul>
   <div class="Navigation__bottom">
 <%- navigationContent() %>
@@ -28,6 +29,7 @@
   <a href="/" class="Navigation__sitelogo"><%= sitedata.siteTitle %></a>
   <ul class="Navigation__main">
     <li><a href="/about/project.html"><%= sitedata.common.layout.navigation.about %></a></li>
+    <li><a href="/about/project.html"><%= sitedata.common.layout.navigation.media %></a></li>
   </ul>
 <%- navigationContent() %>
 </div>
