+++ src/templates/article/detail.ejs
@@ -1,41 +1,41 @@
 <% const dummyImageUrl = 'https://placehold.jp/1280x853.png'; %>
 <% function extend(buf) { %>
-  <article class="MediaArticle">
+  <article class="Article">
 
-    <div class="MediaArticle__photoWrapper">
-      <div class="MediaArticle__photoMain js-MediaArticle__photoMain" style="background-image:url('../../images/mediamock/article_main.jpg')"></div>
+    <div class="Article__photoWrapper">
+      <div class="Article__photoMain js-Article__photoMain" style="background-image:url('../../images/mediamock/article_main.jpg')"></div>
     </div>
 
-    <div class="MediaArticle__inner">
+    <div class="Article__inner">
 
-      <div class="MediaArticle__titleWrapper">
-        <!-- <p class="MediaArticle__titleDate">2017/09/21</p> -->
-        <h1 class="MediaArticle__headingLv1">思い出野郎Aチーム、ダンスフロアの魔法とロマンを信じる男たち</h1>
-        <div class="MediaArticle__titleInfo">
-          <p class="MediaArticle__titleCategory">EVENT</p>
-          <p class="MediaArticle__titleDate">2017.09.21</p>
-          <p class="MediaArticle__titleSeriesName"><a href="TOOD">連載:ほげほげ</a></p>
-          <ul class="MediaArticle__titleTags">
+      <div class="Article__titleWrapper">
+        <!-- <p class="Article__titleDate">2017/09/21</p> -->
+        <h1 class="Article__headingLv1">思い出野郎Aチーム、ダンスフロアの魔法とロマンを信じる男たち</h1>
+        <div class="Article__titleInfo">
+          <p class="Article__titleCategory">EVENT</p>
+          <p class="Article__titleDate">2017.09.21</p>
+          <p class="Article__titleSeriesName"><a href="TOOD">連載:ほげほげ</a></p>
+          <ul class="Article__titleTags">
             <li><a href="TODO">タグ1</a></li>
             <li><a href="TODO">タグ2</a></li>
             <li><a href="TODO">タグ3</a></li>
           </ul>
         </div>
       </div>
-      <h2 class="MediaArticle__headingLv2">プロデューサー・著作権者も救うDubsetとは？</h2>
-      <h3 class="MediaArticle__headingLv3">ーー小杉湯フェスはいかがでしたか</h3>
-      <p class="MediaArticle__text">銭湯でのライブは初めてでしたが、まちがいなく、これまでで一番いい音響でした。建築構造や材料が大きいのかな。建て物は木造で、浴室はタイルですよね。とくに天井が高くてアール型になっているのでよく響くんです。音を反響させるために建てたわけじゃないはずなのに、銭湯と音楽がこんなにも相性がいいなんてものすごい発見ですよね（笑）。この構造のおかげで電気を使う機材が必要ないこともライブの特色として大きい。じつは、ぼくらは2011年に“電気を使わないアコースティックライブ”を下北沢のライブハウスで開催したことがあって。ちょうど3.11が起きた直後でした。そこにお客さんとして来ていたのが、今回の小杉湯フェス企画者の江本祐介くんです。</p>
-      <h3 class="MediaArticle__headingLv3">ーーそのときはライブハウスでの電気なしライブですね。開催のきっかけはなんだったのしょう。</h3>
-      <p class="MediaArticle__text">3.11の後、なんかやらなきゃと思ったんです。とりあえず、ぼくらはミュージシャンだからやっぱりライブしようって。それで、さすがにまだ震災が起きて二、三日くらいだったから節電で「アコギでやろうよ」みたいな軽いノリで。ちょうど「渚」というシングルを出したタイミングで、「渚」のMVが海に入っていく内容だったので、リリースを取り下げたんです。せっかく気合い入れたシングルだったのに悔しい、という気持ちもあったかもしれません。でも、その時の電気なしライブがこうして形を変えてつながっているのはいいですよね。</p>
-      <p class="MediaArticle__text">ほんとうに、歌うことって何もいらないしすごく本来的な行為です。シャムキャッツは歌を大事にするバンドなので、そのスタイルとも合っていたと思います。ギター1本と歌声があればいい。……と、言い切ってしまうとベースとドラムが悲しみますが。（笑） お客さんも「今日はアコースティックセットだから歌がよく聞こえました」と言ってくれて。今後の野望として、銭湯で電気を使ったライブもしてみたい。ベースもドラムも入れて銭湯でバンドやったら絶対いいと思うんです。そういえば、教会でレコーディングするアーティストもいます。銭湯と教会の設計って、すこし似てませんか天井が高くて、光が差し込んで。もう絶対にいい音が出ますよ。あんまり言いすぎるとライブハウスに怒られちゃうけど（笑）。</p>
-      <div class="MediaArticle__innerImage">
+      <h2 class="Article__headingLv2">プロデューサー・著作権者も救うDubsetとは？</h2>
+      <h3 class="Article__headingLv3">ーー小杉湯フェスはいかがでしたか</h3>
+      <p class="Article__text">銭湯でのライブは初めてでしたが、まちがいなく、これまでで一番いい音響でした。建築構造や材料が大きいのかな。建て物は木造で、浴室はタイルですよね。とくに天井が高くてアール型になっているのでよく響くんです。音を反響させるために建てたわけじゃないはずなのに、銭湯と音楽がこんなにも相性がいいなんてものすごい発見ですよね（笑）。この構造のおかげで電気を使う機材が必要ないこともライブの特色として大きい。じつは、ぼくらは2011年に“電気を使わないアコースティックライブ”を下北沢のライブハウスで開催したことがあって。ちょうど3.11が起きた直後でした。そこにお客さんとして来ていたのが、今回の小杉湯フェス企画者の江本祐介くんです。</p>
+      <h3 class="Article__headingLv3">ーーそのときはライブハウスでの電気なしライブですね。開催のきっかけはなんだったのしょう。</h3>
+      <p class="Article__text">3.11の後、なんかやらなきゃと思ったんです。とりあえず、ぼくらはミュージシャンだからやっぱりライブしようって。それで、さすがにまだ震災が起きて二、三日くらいだったから節電で「アコギでやろうよ」みたいな軽いノリで。ちょうど「渚」というシングルを出したタイミングで、「渚」のMVが海に入っていく内容だったので、リリースを取り下げたんです。せっかく気合い入れたシングルだったのに悔しい、という気持ちもあったかもしれません。でも、その時の電気なしライブがこうして形を変えてつながっているのはいいですよね。</p>
+      <p class="Article__text">ほんとうに、歌うことって何もいらないしすごく本来的な行為です。シャムキャッツは歌を大事にするバンドなので、そのスタイルとも合っていたと思います。ギター1本と歌声があればいい。……と、言い切ってしまうとベースとドラムが悲しみますが。（笑） お客さんも「今日はアコースティックセットだから歌がよく聞こえました」と言ってくれて。今後の野望として、銭湯で電気を使ったライブもしてみたい。ベースもドラムも入れて銭湯でバンドやったら絶対いいと思うんです。そういえば、教会でレコーディングするアーティストもいます。銭湯と教会の設計って、すこし似てませんか天井が高くて、光が差し込んで。もう絶対にいい音が出ますよ。あんまり言いすぎるとライブハウスに怒られちゃうけど（笑）。</p>
+      <div class="Article__innerImage">
         <img src="../../images/mediamock/article_inline_1.jpg" alt="">
-        <cite class="MediaArticle__caption">写真のキャプション</cite>
+        <cite class="Article__caption">写真のキャプション</cite>
       </div>
-      <p class="MediaArticle__text">ほんとうに、歌うことって何もいらないしすごく本来的な行為です。シャムキャッツは歌を大事にするバンドなので、そのスタイルとも合っていたと思います。ギター1本と歌声があればいい。……と、言い切ってしまうとベースとドラムが悲しみますが。（笑） お客さんも「今日はアコースティックセットだから歌がよく聞こえました」と言ってくれて。今後の野望として、銭湯で電気を使ったライブもしてみたい。ベースもドラムも入れて銭湯でバンドやったら絶対いいと思うんです。そういえば、教会でレコーディングするアーティストもいます。銭湯と教会の設計って、すこし似てませんか天井が高くて、光が差し込んで。もう絶対にいい音が出ますよ。あんまり言いすぎるとライブハウスに怒られちゃうけど（笑）。</p>
-      <h3 class="MediaArticle__headingLv3">その他の表現例。</h3>
-      <p class="MediaArticle__text"><span style="color: green; font-weight: bold">柏井：</span>閉塞感はめちゃくちゃありましたね。初代「<a href="TODO">CINRA.NET</a>」が2003年からスタートして、法人化したのが2006年。「exPoP!!!!!」が2007年からで、スマホもない、SNSも普及してない時代だったから、同じ思いを抱えている人がいたとしてもそれをどこで共有すればいいのかわからなかった。</p>
-      <p class="MediaArticle__text"><strong>だから「場所をつくる」というのがすごく重要で、</strong>最初は明確に「これをやろう」というよりも、「とりあえず集まろう」という側面が強かったと思います。<strike>その上で、</strike>その時みんなが抱えていた悩みに応える形で、Webメディアやイベントをつくって、そこでみんなの作品を発表しよう、と進んでいきました。
+      <p class="Article__text">ほんとうに、歌うことって何もいらないしすごく本来的な行為です。シャムキャッツは歌を大事にするバンドなので、そのスタイルとも合っていたと思います。ギター1本と歌声があればいい。……と、言い切ってしまうとベースとドラムが悲しみますが。（笑） お客さんも「今日はアコースティックセットだから歌がよく聞こえました」と言ってくれて。今後の野望として、銭湯で電気を使ったライブもしてみたい。ベースもドラムも入れて銭湯でバンドやったら絶対いいと思うんです。そういえば、教会でレコーディングするアーティストもいます。銭湯と教会の設計って、すこし似てませんか天井が高くて、光が差し込んで。もう絶対にいい音が出ますよ。あんまり言いすぎるとライブハウスに怒られちゃうけど（笑）。</p>
+      <h3 class="Article__headingLv3">その他の表現例。</h3>
+      <p class="Article__text"><span style="color: green; font-weight: bold">柏井：</span>閉塞感はめちゃくちゃありましたね。初代「<a href="TODO">CINRA.NET</a>」が2003年からスタートして、法人化したのが2006年。「exPoP!!!!!」が2007年からで、スマホもない、SNSも普及してない時代だったから、同じ思いを抱えている人がいたとしてもそれをどこで共有すればいいのかわからなかった。</p>
+      <p class="Article__text"><strong>だから「場所をつくる」というのがすごく重要で、</strong>最初は明確に「これをやろう」というよりも、「とりあえず集まろう」という側面が強かったと思います。<strike>その上で、</strike>その時みんなが抱えていた悩みに応える形で、Webメディアやイベントをつくって、そこでみんなの作品を発表しよう、と進んでいきました。
 </p>
 
       <div class="Profile">
@@ -72,12 +72,12 @@ const resource = {
   },
   css: [
     '../../styles/common.css',
-    '../../styles/media-article.css'
+    '../../styles/article.css'
   ],
   js: [
     '../../scripts/lib/jquery-3.2.1.min.js',
     '../../scripts/common.bundle.js',
-    '../../scripts/mediaArticle.bundle.js'
+    '../../scripts/article.bundle.js'
   ],
   insertMetaTagFlag: true
 };
