+++ src/templates/article/index.ejs
@@ -25,7 +25,7 @@ const categoryTable = {
 <% function extend(buf) { %>
 <div class="MainVisual" style="background-image:url('../../images/mediamock/top.jpg')">
   <h1 class="MainVisual__title">
-    <div>読む銭湯</div>
+    <div>読みもの</div>
   </h1>
   <!-- <div class="MainVisual__image" style="background-image:url('<%= dummyImageUrl %>')"></div> -->
 </div>
@@ -75,7 +75,7 @@ const resource = {
   },
   css: [
     '../../styles/common.css',
-    '../../styles/media-top.css'
+    '../../styles/article.css'
   ],
   js: [
     '../../scripts/lib/jquery-3.2.1.min.js',
