+++ src/templates/feature/_template.ejs
@@ -1,5 +1,5 @@
 <% function extend(buf) { %>
-<div class="Article">
+<div class="Feature">
 <%
   const imagePath = '../../images/standalone/feature/' + page.id;
   const replaceToBr = str => {return str.replace(/\n/g, '<br>')};
@@ -7,18 +7,18 @@
   const generateText = (textContentArray) => {
     textContentArray.forEach((elm) => {
       if(elm.type === 'headingMain') { %>
-    <div class="Article__headingWrapper">
-      <p class="Article__illust" style="background-image:url('<%= imagePath %>/illust_line.png')"></p>
-      <h1 class="Article__headingLv1"><%= page.title %></h1>
+    <div class="Feature__headingWrapper">
+      <p class="Feature__illust" style="background-image:url('<%= imagePath %>/illust_line.png')"></p>
+      <h1 class="Feature__headingLv1"><%= page.title %></h1>
     </div>
 <% } else if (elm.type === 'headingSub') { %>
-    <h2 class="Article__headingLv2"><%- replaceToBr(elm.content) %></h2>
+    <h2 class="Feature__headingLv2"><%- replaceToBr(elm.content) %></h2>
 <% } else if (elm.type === 'headingAccent') { %>
-    <h3 class="Article__headingLv3"><%- replaceToBr(elm.content) %></h3>
+    <h3 class="Feature__headingLv3"><%- replaceToBr(elm.content) %></h3>
 <% } else if(elm.type === 'text') { %>
-    <p class="Article__text"><%- replaceToBr(elm.content) %></p>
+    <p class="Feature__text"><%- replaceToBr(elm.content) %></p>
 <% } else if(elm.type === 'image') { %>
-    <div class="Article__innerImage">
+    <div class="Feature__innerImage">
       <img src="<%= imagePath + '/' + elm.name %>" alt="">
     </div>
 <% }
@@ -30,11 +30,11 @@
     if(elm.type === 'image') {
       const imageType = (idx === 0) ? 'main' : 'sub';
       const imageName = elm.name ? elm.name : 'photo_' + imageType +'.jpg'; %>
-  <div class="Article__photoWrapper">
-    <div class="Article__photo<%= initialToUpperCase(imageType) %> js-Article__photo<%= initialToUpperCase(imageType) %>" style="background-image:url('<%= imagePath + '/' + imageName %>')"></div>
+  <div class="Feature__photoWrapper">
+    <div class="Feature__photo<%= initialToUpperCase(imageType) %> js-Feature__photo<%= initialToUpperCase(imageType) %>" style="background-image:url('<%= imagePath + '/' + imageName %>')"></div>
   </div>
 <% } else if (elm.type === 'text') { %>
-  <div class="Article__inner">
+  <div class="Feature__inner">
 <%- generateText(elm.content); %>
   </div>
 <% }
@@ -42,7 +42,7 @@
 } %>
 
 <% if(page.profile) { %>
-  <div class="Article__inner">
+  <div class="Feature__inner">
     <div class="Profile">
       <p class="Profile__nameJp"><%= page.profile.nameJp %></p>
       <div class="Profile__nameSub">
@@ -71,12 +71,12 @@ const resource = {
   page: page,
   css: [
     '../../styles/common.css',
-    '../../styles/article.css'
+    '../../styles/feature.css'
   ],
   js: [
     '../../scripts/lib/jquery-3.2.1.min.js',
     '../../scripts/common.bundle.js',
-    '../../scripts/article.bundle.js'
+    '../../scripts/feature.bundle.js'
   ],
   insertMetaTagFlag: true
 };
